Add new feature Create new User/Client
