+++ resources/js/Pages/Auth/RegisterForm.jsx
@@ -1,0 +1,134 @@
+import { useEffect } from 'react';
+import InputError from '@/Components/InputError';
+import InputLabel from '@/Components/InputLabel';
+import PrimaryButton from '@/Components/PrimaryButton';
+import TextInput from '@/Components/TextInput';
+import { Link, useForm } from '@inertiajs/react';
+import ReCAPTCHA from 'react-google-recaptcha';
+
+export default function RegisterForm({ auth, routeCreate = 'register' }) {
+    const { data, setData, post, processing, errors, clearErrors, reset } = useForm({
+        name: '',
+        email: '',
+        password: '',
+        recaptcha_token: '',
+        password_confirmation: '',
+    });
+
+    const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
+    const colorScheme = mediaQuery.matches ? 'dark' : 'light';
+
+    useEffect(() => {
+        return () => {
+            reset('password', 'password_confirmation');
+        };
+    }, []);
+
+    const submit = e => {
+        e.preventDefault();
+
+        post(route(routeCreate));
+    };
+
+    const handleReCaptcha = token => {
+        clearErrors('recaptcha');
+        setData('recaptcha_token', token);
+    }
+
+    return (
+        <form onSubmit={submit}>
+            <div>
+                <InputLabel htmlFor="name" value="Name" />
+
+                <TextInput
+                    id="name"
+                    name="name"
+                    value={data.name}
+                    className="mt-1 block w-full"
+                    autoComplete="name"
+                    isFocused={true}
+                    onChange={(e) => setData('name', e.target.value)}
+                    required
+                />
+
+                <InputError message={errors.name} className="mt-2" />
+            </div>
+
+            <div className="mt-4">
+                <InputLabel htmlFor="email" value="Email" />
+
+                <TextInput
+                    id="email"
+                    type="email"
+                    name="email"
+                    value={data.email}
+                    className="mt-1 block w-full"
+                    autoComplete="username"
+                    onChange={(e) => setData('email', e.target.value)}
+                    required
+                />
+
+                <InputError message={errors.email} className="mt-2" />
+            </div>
+
+            <div className="mt-4">
+                <InputLabel htmlFor="password" value="Password" />
+
+                <TextInput
+                    id="password"
+                    type="password"
+                    name="password"
+                    value={data.password}
+                    className="mt-1 block w-full"
+                    autoComplete="new-password"
+                    onChange={(e) => setData('password', e.target.value)}
+                    required
+                />
+
+                <InputError message={errors.password} className="mt-2" />
+            </div>
+
+            <div className="mt-4">
+                <InputLabel htmlFor="password_confirmation" value="Confirm Password" />
+
+                <TextInput
+                    id="password_confirmation"
+                    type="password"
+                    name="password_confirmation"
+                    value={data.password_confirmation}
+                    className="mt-1 block w-full"
+                    autoComplete="new-password"
+                    onChange={(e) => setData('password_confirmation', e.target.value)}
+                    required
+                />
+
+                <InputError message={errors.password_confirmation} className="mt-2" />
+            </div>
+
+            <div className="mt-4 flex flex-col items-end">
+                <ReCAPTCHA
+                    theme={colorScheme}
+                    sitekey={import.meta.env.VITE_RECAPTCHA_SITE_KEY}
+                    onChange={handleReCaptcha}
+                />
+
+                <InputError message={errors.recaptcha} className="mt-2" />
+            </div>
+
+            <div className="flex items-center justify-end mt-4">
+                {!auth && (
+                    <Link
+                        href={route('login')}
+                        className="underline text-sm text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-100 rounded-md focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 dark:focus:ring-offset-gray-800"
+                    >
+                        Already registered?
+                    </Link>
+                )}
+
+                <PrimaryButton className="ml-4" disabled={processing}>
+                    Register
+                </PrimaryButton>
+            </div>
+        </form>
+    );
+}
